+++ src/whl2conda/impl/prompt.py
@@ -34,7 +34,9 @@ def is_interactive() -> bool:
     """
     True if input appears to be connected to an interactive terminal
     """
-    return sys.__stdin__.isatty()
+    if stdin := sys.__stdin__:
+        return stdin.isatty()
+    return False
 
 
 def bool_input(prompt: str) -> bool: